+++ src/color.ts
@@ -1,0 +1,15 @@
+
+const colors = {
+  error: "red",
+  warn: "orange",
+  info: "blue",
+  debug: "green",
+};
+
+function getColors(): Record<string, string> {
+  return colors;
+}
+
+export {
+  getColors
+}