+++ src/types.ts
@@ -1,7 +1,1 @@
 export type Levels = 'error'|'warn'|'info'|'debug';
-export type Options = {
-  timeZone: string;
-  locale: string;
-  service: string;
-  level: Levels;
-};