+++ src/winstan.d.ts
@@ -8,12 +8,12 @@ export type Options = {
   level: Levels; 
 };
 
-declare function init(options: Options): Logform.Format;
+declare function init(timeZone: string | undefined, locale: string | undefined, service: string | undefined, level: Levels): Logform.Format;
 declare const log: {
-  error: (msg: string, user: string | number, action: string) => void;
-  warn: (msg: string, user: string | number, action: string) => void;
-  info: (msg: string, user: string | number, action: string) => void;
-  debug: (msg: string, user: string | number, action: string) => void;
+    error: (msg: string, id?: string | number | undefined, user?: string | number | undefined, tags?: string[] | number[] | undefined) => void;
+    warn: (msg: string, id?: string | number | undefined, user?: string | number | undefined, tags?: string[] | number[] | undefined) => void;
+    info: (msg: string, id?: string | number | undefined, user?: string | number | undefined, tags?: string[] | number[] | undefined) => void;
+    debug: (msg: string, id?: string | number | undefined, user?: string | number | undefined, tags?: string[] | number[] | undefined) => void;
 };
 
 export {